+++ CHANGELOG.md
@@ -30,6 +30,10 @@
     - You can also configure whether the clip type for the next clip type you create should default to the last clip type you created. This helps with fast creation of multiple clips of the same type. You can enable this default setting in the `SETTINGS > DEFAULTS > UI > CLIP TYPE > USE LAST CLIP TYPE` menu.
     - NOTE for `SONG ROW VIEW`: The default clip type / last clip type setting cannot be used with CV clips and Audio clips in Song Row View. If you set the default to CV or Audio or enable use of the last clip type (and the last clip type is a CV or Audio Clip), then it will create a Synth clip by default. 
 
+#### <ins>Audio Clips</ins>
+- Added audio output modes, and changed audio clip monitoring to be seperate from source selection. Monitoring is now on 
+when the output is a SAMPLER or a LOOPER, chosen by turning the select knob in an audio clip.
+
 #### <ins>Instrument Clip View</ins>
 
 ##### Automation View